SW-4799 Move PlantingSites into scenes/PlantingSitesRouter
- updated import paths
- moved a type into types instead of a circular dependency
