+++ src/scenes/PlantingSitesRouter/PlantingSiteCreate.tsx
@@ -14,9 +14,9 @@ import { useHistory, useParams } from 'react-router-dom';
 import useSnackbar from 'src/utils/useSnackbar';
 import { useAppSelector } from 'src/redux/store';
 import { selectPlantingSite } from 'src/redux/features/tracking/trackingSelectors';
-import PageSnackbar from '../PageSnackbar';
+import PageSnackbar from 'src/components/PageSnackbar';
 import { PlantingSite, UpdatedPlantingSeason } from 'src/types/Tracking';
-import BoundariesAndZones from 'src/components/PlantingSites/BoundariesAndZones';
+import BoundariesAndZones from './BoundariesAndZones';
 import { useOrganization } from 'src/providers/hooks';
 import Card from 'src/components/common/Card';
 import PlantingSiteMapEditor from 'src/components/Map/PlantingSiteMapEditor';
