+++ src/scenes/PlantingSitesRouter/PlantingSiteTypeSelect.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-dom';
 import { APP_PATHS } from 'src/constants';
 import { SiteType } from 'src/types/PlantingSite';
 import isEnabled from 'src/features';
-import DetailedPlantingSiteHelpModal from 'src/components/PlantingSites/DetailedPlantingSiteHelpModal';
+import DetailedPlantingSiteHelpModal from './DetailedPlantingSiteHelpModal';
 import PlantingSiteSelectTypeModal from './PlantingSiteSelectTypeModal';
 import PlantingSiteSelectTypeModal2 from './PlantingSiteSelectTypeModal2';
