+++ src/scenes/PlantingSitesRouter/PlantingSitesTable.tsx
@@ -1,18 +1,17 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { Box, Grid, useTheme } from '@mui/material';
 import { PillList, PillListItem, TableColumnType, Textfield } from '@terraware/web-components';
-import { SearchResponseElement } from 'src/types/Search';
+import { SearchResponseElement, SearchSortOrder } from 'src/types/Search';
 import strings from 'src/strings';
 import PlantingSitesCellRenderer from './PlantingSitesCellRenderer';
 import Table from 'src/components/common/table';
 import { SortOrder } from 'src/components/common/table/sort';
-import { SearchSortOrder } from 'src/types/Search';
 import isEnabled from 'src/features';
 import FilterMultiSelectContainer from 'src/components/common/FilterMultiSelectContainer';
 import { selectProjects } from 'src/redux/features/projects/projectsSelectors';
 import { useAppSelector } from 'src/redux/store';
 import { Project } from 'src/types/Project';
-import { PlantingSitesFilters } from 'src/components/PlantingSites/PlantingSitesList';
+import { PlantingSitesFilters } from 'src/types/PlantingSite';
 import Card from 'src/components/common/Card';
 
 interface PlantingSitesTableProps {
